feat: connect faucet/approve to eth tx store
